Update
Code cleanup for compilation with Clang and MSVC
Add user service layer
Add an option to disable default key behavior
Disable notifier surface from rendering
Instanciable Tracer
Refactor Tracer initilization with logger to use std location
Update primary services identification
Improve Window fullscreen management
Adds C++ constants instead of macros
Add the visibility check per surfaces for input in overlay manager.

+++ src/Saphir/Generator/SceneRendering.cpp
@@ -42,8 +42,13 @@ namespace EmEn::Saphir::Generator
 	using namespace Vulkan;
 	using namespace Keys;
 
-	SceneRendering::SceneRendering (Settings & settings, const std::string & name, const std::shared_ptr< const RenderTarget::Abstract > & renderTarget, const std::shared_ptr< const RenderableInstance::Abstract > & renderableInstance, size_t layerIndex, RenderPassType renderPassType, const Scenes::Scene & scene) noexcept
-		: Abstract(settings, name, renderTarget), m_renderPassType(renderPassType), m_lightGenerator(renderPassType), m_renderableInstance(renderableInstance), m_layerIndex(layerIndex), m_scene(&scene)
+	SceneRendering::SceneRendering (Settings & settings, const std::string & name, const std::shared_ptr< const RenderTarget::Abstract > & renderTarget, const std::shared_ptr< const RenderableInstance::Abstract > & renderableInstance, uint32_t layerIndex, RenderPassType renderPassType, const Scenes::Scene & scene) noexcept
+		: Abstract(settings, name, renderTarget),
+		m_renderPassType(renderPassType),
+		m_lightGenerator(renderPassType),
+		m_renderableInstance(renderableInstance),
+		m_layerIndex(layerIndex),
+		m_scene(&scene)
 	{
 
 	}
